# full-stack-todo/specs/1-backend-api-persistence/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: |
    RESTful API for task management with JWT authentication and user data isolation.

    ## Authentication
    All endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## User Isolation
    All operations are automatically scoped to the authenticated user. Users can only
    access, modify, or delete their own tasks.

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

security:
  - BearerAuth: []

tags:
  - name: tasks
    description: Task management operations
  - name: health
    description: Health check endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Returns the health status of the API
      tags:
        - health
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/tasks:
    get:
      summary: List all tasks
      description: Returns all tasks belonging to the authenticated user, ordered by creation date (newest first)
      tags:
        - tasks
      responses:
        '200':
          description: Successful response with array of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          minimum: 1

    get:
      summary: Get a specific task
      description: Returns a specific task if it belongs to the authenticated user
      tags:
        - tasks
      responses:
        '200':
          description: Successful response with task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update a task
      description: Updates a task if it belongs to the authenticated user
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete a task
      description: Deletes a task if it belongs to the authenticated user
      tags:
        - tasks
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/tasks/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          minimum: 1

    patch:
      summary: Toggle task completion
      description: Toggles the completion status of a task (true ↔ false)
      tags:
        - tasks
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication system

  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        title:
          type: string
          maxLength: 255
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Detailed task description
          example: "Write comprehensive API documentation"
        completed:
          type: boolean
          description: Task completion status
          example: false
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 123
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
          example: "2026-02-08T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated
          example: "2026-02-08T10:30:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required, non-empty)
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Detailed task description (optional)
          example: "Write comprehensive API documentation"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title
          example: "Updated task title"
        description:
          type: string
          nullable: true
          description: Updated task description
          example: "Updated description"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    Error:
      type: object
      required:
        - detail
        - status_code
        - timestamp
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Task not found"
        status_code:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Timestamp when error occurred
          example: "2026-02-08T10:30:00Z"

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            emptyTitle:
              summary: Empty title
              value:
                detail: "Title cannot be empty"
                status_code: 400
                timestamp: "2026-02-08T10:30:00Z"
            titleTooLong:
              summary: Title too long
              value:
                detail: "Title must be 255 characters or less"
                status_code: 400
                timestamp: "2026-02-08T10:30:00Z"

    Unauthorized:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing token
              value:
                detail: "Authorization header missing"
                status_code: 401
                timestamp: "2026-02-08T10:30:00Z"
            invalidToken:
              summary: Invalid token
              value:
                detail: "Invalid JWT token"
                status_code: 401
                timestamp: "2026-02-08T10:30:00Z"
            expiredToken:
              summary: Expired token
              value:
                detail: "JWT token has expired"
                status_code: 401
                timestamp: "2026-02-08T10:30:00Z"

    Forbidden:
      description: Valid JWT but insufficient permissions (task belongs to different user)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authorized to access this task"
            status_code: 403
            timestamp: "2026-02-08T10:30:00Z"

    NotFound:
      description: Task not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"
            status_code: 404
            timestamp: "2026-02-08T10:30:00Z"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"
            status_code: 500
            timestamp: "2026-02-08T10:30:00Z"